server: split media source providers
