+++ crates/server/src/source.rs
@@ -29,14 +29,23 @@ use async_trait::async_trait;
 use config::{AppConfig, MusicService, Source};
 use db::Db;
 
-use crate::jellyfin::JellyfinClient;
 use crate::server_ops::ServerConn;
-use crate::subsonic::SubsonicClient;
-use crate::ytmusic::YouTubeMusicClient;
 
 pub mod capabilities;
+mod jellyfin;
+mod local;
+mod offline;
+mod soundcloud;
+mod subsonic;
 mod types;
+mod youtube_music;
+use jellyfin::JellyfinSource;
+use local::LocalSource;
+use offline::OfflineServerSource;
+use soundcloud::SoundcloudSource;
+use subsonic::SubsonicSource;
 pub use types::*;
+use youtube_music::YtSource;
 
 /// The source-agnostic backend the app drives. Impls supply
 /// [`source`](Self::source) + [`db`](Self::db) (the seam the default methods
@@ -658,7 +667,7 @@ pub trait MediaSource: Send + Sync {
 
 /// Mirror a successful remote playlist-add into the DB cache (so the DB-reading
 /// UI reflects it without waiting for a sync).
-async fn mirror_added(
+pub(super) async fn mirror_added(
     db: &Db,
     source: &Source,
     pid: &str,
@@ -672,7 +681,7 @@ async fn mirror_added(
 
 /// Encode a cover URL into the `urlhex_…` tag the Subsonic cover seam decodes
 /// back (the synthetic album/track cover reference for Subsonic/Custom).
-fn encode_cover_url_tag(url: &str) -> String {
+pub(super) fn encode_cover_url_tag(url: &str) -> String {
     let mut hex = String::with_capacity(url.len() * 2);
     for b in url.as_bytes() {
         hex.push_str(&format!("{b:02x}"));
@@ -724,7 +733,7 @@ fn search_filter(
 }
 
 /// Mirror a remote playlist-create into the DB cache.
-async fn mirror_created(
+pub(super) async fn mirror_created(
     db: &Db,
     source: &Source,
     id: &str,
@@ -736,1228 +745,6 @@ async fn mirror_created(
     db.set_playlist_tracks(source, id, refs)
         .await
         .map_err(SourceError::from)
-}
-
-// ============================ Local ====================================
-
-/// The local (filesystem) source: every op is a DB write or a no-op.
-struct LocalSource {
-    db: Db,
-    source: Source,
-}
-
-#[async_trait]
-impl MediaSource for LocalSource {
-    fn source(&self) -> &Source {
-        &self.source
-    }
-    fn db(&self) -> &Db {
-        &self.db
-    }
-
-    fn capabilities(&self) -> Capabilities {
-        Capabilities {
-            edit_tags: true,
-            delete_from_disk: true,
-            scan_folders: true,
-            folders: true,
-            sync: false,
-            downloads: false,
-            discover: false,
-            radio: false,
-            playlists: PlaylistOps::Reorder,
-            artist_view: ArtistView::Library,
-            albums: AlbumType::Standard,
-            favorites_sync: FavoritesSync::Instant,
-        }
-    }
-
-    async fn add_to_playlist(
-        &self,
-        playlist_id: &str,
-        item_refs: &[String],
-    ) -> Result<Vec<String>, SourceError> {
-        self.db
-            .add_playlist_tracks(&self.source, playlist_id, item_refs)
-            .await?;
-        Ok(item_refs.to_vec())
-    }
-
-    async fn create_playlist(
-        &self,
-        name: &str,
-        item_refs: &[String],
-    ) -> Result<String, SourceError> {
-        let id = uuid::Uuid::new_v4().to_string();
-        mirror_created(&self.db, &self.source, &id, name, item_refs).await?;
-        Ok(id)
-    }
-
-    async fn remove_from_playlist(
-        &self,
-        playlist_id: &str,
-        track: &reader::Track,
-        _position: usize,
-    ) -> Result<(), SourceError> {
-        let r = track.id.key().into_owned();
-        self.db
-            .remove_playlist_tracks(&self.source, playlist_id, &[r])
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn reorder_playlist(
-        &self,
-        playlist_id: &str,
-        ordered_refs: &[String],
-        _moved: &reader::Track,
-        _new_index: usize,
-    ) -> Result<(), SourceError> {
-        self.db
-            .set_playlist_tracks(&self.source, playlist_id, ordered_refs)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn resolve_stream(&self, item_id: &str) -> Result<StreamInfo, SourceError> {
-        Ok(StreamInfo {
-            url: item_id.to_string(),
-            format: None,
-            user_agent: None,
-            duration_secs: None,
-            bitrate: None,
-            content_length: None,
-        })
-    }
-
-    async fn validate(&self) -> AuthOutcome {
-        AuthOutcome::Valid
-    }
-
-    async fn fetch_favorites(&self) -> Result<Vec<String>, SourceError> {
-        Ok(Vec::new())
-    }
-
-    async fn push_favorite(&self, _item_id: &str, _on: bool) -> Result<(), SourceError> {
-        Ok(())
-    }
-}
-
-// ===================== Server, no live connection ======================
-
-/// A server whose creds are missing/unusable: favorites + cache mutators still
-/// work (DB defaults), every remote op errors. Keeps anonymous likes queuing.
-struct OfflineServerSource {
-    db: Db,
-    source: Source,
-}
-
-#[async_trait]
-impl MediaSource for OfflineServerSource {
-    fn source(&self) -> &Source {
-        &self.source
-    }
-    fn db(&self) -> &Db {
-        &self.db
-    }
-
-    fn capabilities(&self) -> Capabilities {
-        // Creds-less / unreachable: no remote affordances. Favorites still queue
-        // (a default DB op, not gated here).
-        Capabilities {
-            edit_tags: false,
-            delete_from_disk: false,
-            scan_folders: false,
-            folders: false,
-            sync: false,
-            downloads: false,
-            discover: false,
-            radio: false,
-            playlists: PlaylistOps::None,
-            artist_view: ArtistView::Library,
-            albums: AlbumType::Standard,
-            favorites_sync: FavoritesSync::Instant,
-        }
-    }
-
-    async fn add_to_playlist(&self, _: &str, _: &[String]) -> Result<Vec<String>, SourceError> {
-        Err(SourceError::Connectivity)
-    }
-    async fn create_playlist(&self, _: &str, _: &[String]) -> Result<String, SourceError> {
-        Err(SourceError::Connectivity)
-    }
-    async fn remove_from_playlist(
-        &self,
-        _: &str,
-        _: &reader::Track,
-        _: usize,
-    ) -> Result<(), SourceError> {
-        Err(SourceError::Connectivity)
-    }
-    async fn resolve_stream(&self, _: &str) -> Result<StreamInfo, SourceError> {
-        Err(SourceError::Auth)
-    }
-    async fn validate(&self) -> AuthOutcome {
-        AuthOutcome::Unreachable
-    }
-    async fn fetch_favorites(&self) -> Result<Vec<String>, SourceError> {
-        Err(SourceError::Connectivity)
-    }
-    async fn push_favorite(&self, _: &str, _: bool) -> Result<(), SourceError> {
-        Err(SourceError::Connectivity)
-    }
-}
-
-// ============================ Jellyfin =================================
-
-struct JellyfinSource {
-    db: Db,
-    source: Source,
-    client: JellyfinClient,
-}
-
-#[async_trait]
-impl MediaSource for JellyfinSource {
-    fn source(&self) -> &Source {
-        &self.source
-    }
-    fn db(&self) -> &Db {
-        &self.db
-    }
-
-    async fn keepalive(&self) -> Result<(), SourceError> {
-        self.client.ping().await.map_err(SourceError::from)
-    }
-
-    async fn report_playback_start(&self, item_id: &str) -> Result<(), SourceError> {
-        self.client
-            .report_playback_start(item_id)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn report_playback_stopped(
-        &self,
-        item_id: &str,
-        position_ticks: u64,
-    ) -> Result<(), SourceError> {
-        self.client
-            .report_playback_stopped(item_id, position_ticks)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn report_playback_progress(
-        &self,
-        item_id: &str,
-        position_ticks: u64,
-        is_paused: bool,
-    ) -> Result<(), SourceError> {
-        self.client
-            .report_playback_progress(item_id, position_ticks, is_paused)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_library(&self) -> Result<LibrarySnapshot, SourceError> {
-        use std::path::PathBuf;
-        let mut albums = Vec::new();
-        let mut tracks = Vec::new();
-        let mut artist_images = Vec::new();
-
-        for lib in self.client.get_music_libraries().await? {
-            let mut start = 0;
-            let limit = 500;
-            loop {
-                let (page, _total) = self
-                    .client
-                    .get_albums_paginated(&lib.id, start, limit)
-                    .await?;
-                if page.is_empty() {
-                    break;
-                }
-                let count = page.len();
-                for a in page {
-                    let image_tag = a
-                        .image_tags
-                        .as_ref()
-                        .and_then(|t| t.get("Primary").cloned());
-                    let cover_path = Some(PathBuf::from(match &image_tag {
-                        Some(tag) => format!("jellyfin:{}:{}", a.id, tag),
-                        None => format!("jellyfin:{}", a.id),
-                    }));
-                    albums.push(reader::Album {
-                        id: format!("jellyfin:{}", a.id),
-                        title: a.name,
-                        artist: a
-                            .album_artist
-                            .or_else(|| a.artists.as_ref().map(|x| x.join(", ")))
-                            .unwrap_or_default(),
-                        genre: a.genres.as_ref().map(|g| g.join(", ")).unwrap_or_default(),
-                        year: a
-                            .production_year
-                            .map(|y| u16::try_from(y).unwrap_or(u16::MAX))
-                            .unwrap_or(0),
-                        cover_path,
-                        manual_cover: false,
-                    });
-                }
-                start += count;
-                if count < limit {
-                    break;
-                }
-            }
-
-            let mut start = 0;
-            let limit = 500;
-            loop {
-                let items = self
-                    .client
-                    .get_music_library_items_paginated(&lib.id, start, limit)
-                    .await?;
-                if items.is_empty() {
-                    break;
-                }
-                let count = items.len();
-                for item in items {
-                    let cover = item
-                        .image_tags
-                        .as_ref()
-                        .and_then(|tags| tags.get("Primary").cloned());
-                    let bitrate_u16 =
-                        (item.bitrate.unwrap_or(0) / 1000).min(u16::MAX as u32) as u16;
-                    tracks.push(reader::Track {
-                        id: reader::models::TrackId::Server {
-                            service: MusicService::Jellyfin,
-                            item_id: item.id.clone(),
-                        },
-                        cover,
-                        album_id: item
-                            .album_id
-                            .map(|id| format!("jellyfin:{}", id))
-                            .unwrap_or_default(),
-                        title: item.name,
-                        artist: item
-                            .album_artist
-                            .clone()
-                            .or_else(|| item.artists.as_ref().map(|a| a.join(", ")))
-                            .unwrap_or_default(),
-                        album: item.album.unwrap_or_default(),
-                        duration: item.run_time_ticks.unwrap_or(0) / 10_000_000,
-                        khz: item.sample_rate.unwrap_or(0),
-                        bitrate: bitrate_u16,
-                        track_number: item.index_number,
-                        disc_number: item.parent_index_number,
-                        musicbrainz_release_id: None,
-                        musicbrainz_recording_id: None,
-                        musicbrainz_track_id: None,
-                        playlist_item_id: None,
-                        artists: item
-                            .artists
-                            .unwrap_or_else(|| item.album_artist.into_iter().collect()),
-                    });
-                }
-                start += count;
-                if count < limit {
-                    break;
-                }
-            }
-        }
-
-        if let Ok(artists) = self.client.get_artists().await {
-            for artist in artists {
-                if let Some(tags) = &artist.image_tags
-                    && let Some(tag) = tags.get("Primary")
-                {
-                    let url = utils::jellyfin_image::jellyfin_image_url(
-                        self.client.base_url(),
-                        &artist.id,
-                        Some(tag.as_str()),
-                        self.client.token(),
-                        512,
-                        90,
-                    );
-                    artist_images.push((artist.name, url));
-                }
-            }
-        }
-
-        Ok(LibrarySnapshot {
-            albums,
-            tracks,
-            artist_images,
-        })
-    }
-
-    async fn set_playlist_cover(
-        &self,
-        playlist_id: &str,
-        name: &str,
-        image_path: &std::path::Path,
-        image_tag: Option<&str>,
-    ) -> Result<(), SourceError> {
-        // Push the artwork to Jellyfin (best-effort), then record it locally.
-        if let Ok(bytes) = std::fs::read(image_path) {
-            let ct = match image_path
-                .extension()
-                .and_then(|e| e.to_str())
-                .map(str::to_lowercase)
-                .as_deref()
-            {
-                Some("png") => "image/png",
-                _ => "image/jpeg",
-            };
-            let _ = self.client.set_playlist_image(playlist_id, bytes, ct).await;
-        }
-        let cover = image_path.to_string_lossy();
-        self.db()
-            .upsert_playlist_meta(self.source(), playlist_id, name, Some(&cover), image_tag)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    fn capabilities(&self) -> Capabilities {
-        Capabilities {
-            edit_tags: false,
-            delete_from_disk: false,
-            scan_folders: false,
-            folders: false,
-            sync: true,
-            downloads: true,
-            discover: false,
-            radio: false,
-            playlists: PlaylistOps::Reorder,
-            artist_view: ArtistView::Library,
-            albums: AlbumType::Standard,
-            favorites_sync: FavoritesSync::Instant,
-        }
-    }
-
-    async fn add_to_playlist(
-        &self,
-        playlist_id: &str,
-        item_refs: &[String],
-    ) -> Result<Vec<String>, SourceError> {
-        let mut added = Vec::new();
-        for id in item_refs {
-            if self.client.add_to_playlist(playlist_id, id).await.is_ok() {
-                added.push(id.clone());
-            }
-        }
-        mirror_added(&self.db, &self.source, playlist_id, &added).await?;
-        Ok(added)
-    }
-
-    async fn create_playlist(
-        &self,
-        name: &str,
-        item_refs: &[String],
-    ) -> Result<String, SourceError> {
-        let refs: Vec<&str> = item_refs.iter().map(String::as_str).collect();
-        let id = self.client.create_playlist(name, &refs).await?;
-        mirror_created(&self.db, &self.source, &id, name, item_refs).await?;
-        Ok(id)
-    }
-
-    async fn remove_from_playlist(
-        &self,
-        playlist_id: &str,
-        track: &reader::Track,
-        _position: usize,
-    ) -> Result<(), SourceError> {
-        let entry_id = track
-            .playlist_item_id
-            .as_deref()
-            .ok_or_else(|| SourceError::InvalidInput("track has no playlist-entry id".into()))?;
-        self.client
-            .remove_from_playlist(playlist_id, entry_id)
-            .await?;
-        self.db
-            .remove_playlist_tracks(&self.source, playlist_id, &[track.id.key().into_owned()])
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn resolve_stream(&self, item_id: &str) -> Result<StreamInfo, SourceError> {
-        Ok(StreamInfo {
-            url: self.client.stream_url(item_id),
-            format: None,
-            user_agent: None,
-            duration_secs: None,
-            bitrate: None,
-            content_length: None,
-        })
-    }
-
-    async fn validate(&self) -> AuthOutcome {
-        match self.client.ping().await {
-            Ok(()) => AuthOutcome::Valid,
-            Err(e) if e.contains("401") || e.contains("403") => AuthOutcome::Expired,
-            Err(_) => AuthOutcome::Unreachable,
-        }
-    }
-
-    async fn fetch_favorites(&self) -> Result<Vec<String>, SourceError> {
-        Ok(self
-            .client
-            .get_favorite_items()
-            .await?
-            .into_iter()
-            .map(|i| i.id)
-            .collect())
-    }
-
-    async fn push_favorite(&self, item_id: &str, on: bool) -> Result<(), SourceError> {
-        if on {
-            self.client.mark_favorite(item_id).await
-        } else {
-            self.client.unmark_favorite(item_id).await
-        }
-        .map_err(SourceError::from)
-    }
-
-    async fn reorder_playlist(
-        &self,
-        playlist_id: &str,
-        ordered_refs: &[String],
-        moved: &reader::Track,
-        new_index: usize,
-    ) -> Result<(), SourceError> {
-        let entry_id = moved
-            .playlist_item_id
-            .as_deref()
-            .ok_or_else(|| SourceError::InvalidInput("track has no playlist-entry id".into()))?;
-        self.client
-            .move_playlist_item(playlist_id, entry_id, new_index)
-            .await?;
-        self.db
-            .set_playlist_tracks(&self.source, playlist_id, ordered_refs)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_playlists(&self) -> Result<Vec<PlaylistMeta>, SourceError> {
-        Ok(self
-            .client
-            .get_playlists()
-            .await?
-            .into_iter()
-            .map(|p| PlaylistMeta {
-                id: p.id,
-                name: p.name,
-                image_tag: p
-                    .image_tags
-                    .as_ref()
-                    .and_then(|tags| tags.get("Primary").cloned()),
-            })
-            .collect())
-    }
-
-    async fn fetch_playlist_entries(
-        &self,
-        playlist_id: &str,
-    ) -> Result<Vec<reader::Track>, SourceError> {
-        let items = self.client.get_playlist_items(playlist_id).await?;
-        Ok(items
-            .into_iter()
-            .map(|item| {
-                let duration_secs = item.run_time_ticks.unwrap_or(0) / 10_000_000;
-                let cover = item
-                    .image_tags
-                    .as_ref()
-                    .and_then(|tags| tags.get("Primary").cloned());
-                let bitrate_kbps = item.bitrate.unwrap_or(0) / 1000;
-                let artist_str = item
-                    .album_artist
-                    .clone()
-                    .or_else(|| item.artists.as_ref().map(|a| a.join(", ")))
-                    .unwrap_or_default();
-                reader::models::Track {
-                    id: reader::models::TrackId::Server {
-                        service: MusicService::Jellyfin,
-                        item_id: item.id.clone(),
-                    },
-                    cover,
-                    album_id: item
-                        .album_id
-                        .map(|id| format!("jellyfin:{}", id))
-                        .unwrap_or_default(),
-                    title: item.name,
-                    artist: artist_str,
-                    album: item.album.unwrap_or_default(),
-                    duration: duration_secs,
-                    khz: item.sample_rate.unwrap_or(0),
-                    bitrate: bitrate_kbps.min(u16::MAX as u32) as u16,
-                    track_number: item.index_number,
-                    disc_number: item.parent_index_number,
-                    musicbrainz_release_id: None,
-                    musicbrainz_recording_id: None,
-                    musicbrainz_track_id: None,
-                    playlist_item_id: item.playlist_item_id,
-                    artists: item.artists.unwrap_or_default(),
-                }
-            })
-            .collect())
-    }
-
-    async fn fetch_artist_images(&self) -> Result<Vec<(String, String)>, SourceError> {
-        let artists = self.client.get_artists().await?;
-        let mut out = Vec::new();
-        for artist in artists {
-            if let Some(tags) = &artist.image_tags
-                && let Some(tag) = tags.get("Primary")
-            {
-                out.push((
-                    artist.name.clone(),
-                    utils::jellyfin_image::jellyfin_image_url(
-                        self.client.base_url(),
-                        &artist.id,
-                        Some(tag.as_str()),
-                        self.client.token(),
-                        512,
-                        90,
-                    ),
-                ));
-            }
-        }
-        Ok(out)
-    }
-}
-
-// ============================ Subsonic =================================
-
-struct SubsonicSource {
-    db: Db,
-    source: Source,
-    client: SubsonicClient,
-    /// Subsonic or Custom — both use this impl; the typed track id needs the
-    /// exact one so its uid prefix round-trips.
-    service: MusicService,
-}
-
-#[async_trait]
-impl MediaSource for SubsonicSource {
-    fn source(&self) -> &Source {
-        &self.source
-    }
-    fn db(&self) -> &Db {
-        &self.db
-    }
-
-    async fn scrobble_now_playing(&self, item_id: &str) -> Result<(), SourceError> {
-        self.client
-            .scrobble_now_playing(item_id)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn scrobble(&self, item_id: &str) -> Result<(), SourceError> {
-        self.client
-            .scrobble(item_id)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_library(&self) -> Result<LibrarySnapshot, SourceError> {
-        use std::path::PathBuf;
-        let prefix = match self.service {
-            MusicService::Custom => "custom",
-            _ => "subsonic",
-        };
-        let mut albums = Vec::new();
-        let mut tracks = Vec::new();
-        let mut artist_images = Vec::new();
-        let mut seen = std::collections::HashSet::new();
-
-        if let Ok(artists) = self.client.get_artists().await {
-            for artist in artists {
-                if let Some(cover_art_id) = &artist.cover_art
-                    && let Ok(url) = self.client.cover_art_url(cover_art_id, Some(512))
-                {
-                    artist_images.push((artist.name, url));
-                }
-            }
-        }
-
-        let mut offset = 0;
-        let batch = 250;
-        loop {
-            let page = self.client.get_album_list(offset, batch).await?;
-            if page.is_empty() {
-                break;
-            }
-            let count = page.len();
-            for album in page {
-                let album_cover_tag = album
-                    .cover_art
-                    .as_ref()
-                    .and_then(|c| self.client.cover_art_url(c, Some(512)).ok())
-                    .map(|url| encode_cover_url_tag(&url));
-                let album_id_prefixed = match &album_cover_tag {
-                    Some(tag) => format!("{}:{}:{}", prefix, album.id, tag),
-                    None => format!("{}:{}:none", prefix, album.id),
-                };
-                let album_name = album.name.clone();
-                let album_artist = album.artist.clone().unwrap_or_default();
-                albums.push(reader::Album {
-                    id: album_id_prefixed.clone(),
-                    title: album_name.clone(),
-                    artist: album_artist.clone(),
-                    genre: album.genre.clone().unwrap_or_default(),
-                    year: album.year.unwrap_or(0),
-                    cover_path: Some(PathBuf::from(album_id_prefixed.clone())),
-                    manual_cover: false,
-                });
-
-                let songs = self.client.get_album_songs(&album.id).await.map_err(|e| {
-                    SourceError::Backend(format!(
-                        "failed to fetch songs for album {}: {e}",
-                        album.id
-                    ))
-                })?;
-                for song in songs {
-                    if !seen.insert(song.id.clone()) {
-                        continue;
-                    }
-                    let bitrate_u16 = song.bit_rate.unwrap_or(0).min(u16::MAX as u32) as u16;
-                    let song_cover_tag = song
-                        .cover_art
-                        .as_ref()
-                        .and_then(|c| self.client.cover_art_url(c, Some(512)).ok())
-                        .map(|url| encode_cover_url_tag(&url));
-                    tracks.push(reader::Track {
-                        id: reader::models::TrackId::Server {
-                            service: self.service,
-                            item_id: song.id.clone(),
-                        },
-                        cover: Some(song_cover_tag.unwrap_or_else(|| "none".to_string())),
-                        album_id: album_id_prefixed.clone(),
-                        title: song.title,
-                        artist: song.artist.clone().unwrap_or_else(|| album_artist.clone()),
-                        album: song.album.unwrap_or_else(|| album_name.clone()),
-                        duration: song.duration.unwrap_or(0),
-                        khz: song.sampling_rate.unwrap_or(0),
-                        bitrate: bitrate_u16,
-                        track_number: song.track,
-                        disc_number: song.disc_number,
-                        musicbrainz_release_id: None,
-                        musicbrainz_recording_id: None,
-                        musicbrainz_track_id: None,
-                        playlist_item_id: None,
-                        artists: vec![song.artist.unwrap_or_else(|| album_artist.clone())],
-                    });
-                }
-            }
-            offset += count;
-            if count < batch {
-                break;
-            }
-        }
-
-        Ok(LibrarySnapshot {
-            albums,
-            tracks,
-            artist_images,
-        })
-    }
-
-    fn capabilities(&self) -> Capabilities {
-        Capabilities {
-            edit_tags: false,
-            delete_from_disk: false,
-            scan_folders: false,
-            folders: false,
-            sync: true,
-            downloads: true,
-            discover: false,
-            radio: false,
-            playlists: PlaylistOps::Reorder,
-            artist_view: ArtistView::Library,
-            albums: AlbumType::Standard,
-            favorites_sync: FavoritesSync::Instant,
-        }
-    }
-
-    async fn add_to_playlist(
-        &self,
-        playlist_id: &str,
-        item_refs: &[String],
-    ) -> Result<Vec<String>, SourceError> {
-        let mut added = Vec::new();
-        for id in item_refs {
-            if self.client.add_to_playlist(playlist_id, id).await.is_ok() {
-                added.push(id.clone());
-            }
-        }
-        mirror_added(&self.db, &self.source, playlist_id, &added).await?;
-        Ok(added)
-    }
-
-    async fn create_playlist(
-        &self,
-        name: &str,
-        item_refs: &[String],
-    ) -> Result<String, SourceError> {
-        let refs: Vec<&str> = item_refs.iter().map(String::as_str).collect();
-        let id = self.client.create_playlist(name, &refs).await?;
-        mirror_created(&self.db, &self.source, &id, name, item_refs).await?;
-        Ok(id)
-    }
-
-    async fn remove_from_playlist(
-        &self,
-        playlist_id: &str,
-        track: &reader::Track,
-        position: usize,
-    ) -> Result<(), SourceError> {
-        self.client
-            .remove_from_playlist(playlist_id, position)
-            .await?;
-        self.db
-            .remove_playlist_tracks(&self.source, playlist_id, &[track.id.key().into_owned()])
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn resolve_stream(&self, item_id: &str) -> Result<StreamInfo, SourceError> {
-        Ok(StreamInfo {
-            url: self.client.stream_url(item_id)?,
-            format: None,
-            user_agent: None,
-            duration_secs: None,
-            bitrate: None,
-            content_length: None,
-        })
-    }
-
-    async fn validate(&self) -> AuthOutcome {
-        match self.client.ping().await {
-            Ok(()) => AuthOutcome::Valid,
-            Err(e)
-                if e.contains("Wrong username")
-                    || e.contains("not authorized")
-                    || e.contains("code 40") =>
-            {
-                AuthOutcome::Expired
-            }
-            Err(_) => AuthOutcome::Unreachable,
-        }
-    }
-
-    async fn fetch_favorites(&self) -> Result<Vec<String>, SourceError> {
-        self.client
-            .get_starred_song_ids()
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn push_favorite(&self, item_id: &str, on: bool) -> Result<(), SourceError> {
-        if on {
-            self.client.star(item_id).await
-        } else {
-            self.client.unstar(item_id).await
-        }
-        .map_err(SourceError::from)
-    }
-
-    async fn reorder_playlist(
-        &self,
-        playlist_id: &str,
-        ordered_refs: &[String],
-        _moved: &reader::Track,
-        _new_index: usize,
-    ) -> Result<(), SourceError> {
-        let ids: Vec<&str> = ordered_refs.iter().map(String::as_str).collect();
-        self.client
-            .reorder_playlist(playlist_id, &ids, ids.len())
-            .await?;
-        self.db
-            .set_playlist_tracks(&self.source, playlist_id, ordered_refs)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_playlists(&self) -> Result<Vec<PlaylistMeta>, SourceError> {
-        Ok(self
-            .client
-            .get_playlists()
-            .await?
-            .into_iter()
-            .map(|p| PlaylistMeta {
-                id: p.id,
-                name: p.name,
-                image_tag: None,
-            })
-            .collect())
-    }
-
-    async fn fetch_playlist_entries(
-        &self,
-        playlist_id: &str,
-    ) -> Result<Vec<reader::Track>, SourceError> {
-        let items = self.client.get_playlist_entries(playlist_id).await?;
-        Ok(items
-            .into_iter()
-            .map(|item| {
-                // Encode the cover URL as the `urlhex_` tag the cover resolver
-                // understands; album_id carries it (or `:none`) like subsonic_sync.
-                let cover_tag = item
-                    .cover_art
-                    .as_ref()
-                    .and_then(|id| self.client.cover_art_url(id, Some(512)).ok())
-                    .map(|url| format!("urlhex_{}", hex::encode(url.as_bytes())));
-                let album_id = item
-                    .album_id
-                    .as_ref()
-                    .map(|id| match &cover_tag {
-                        Some(tag) => format!("jellyfin:{}:{}", id, tag),
-                        None => format!("jellyfin:{}:none", id),
-                    })
-                    .unwrap_or_else(|| format!("jellyfin:{}:none", item.id));
-                let artist = item.artist.clone().unwrap_or_default();
-                reader::models::Track {
-                    id: reader::models::TrackId::Server {
-                        service: self.service,
-                        item_id: item.id.clone(),
-                    },
-                    cover: Some(cover_tag.unwrap_or_else(|| "none".to_string())),
-                    album_id,
-                    title: item.title,
-                    artist: artist.clone(),
-                    album: item.album.unwrap_or_default(),
-                    duration: item.duration.unwrap_or(0),
-                    khz: item.sampling_rate.unwrap_or(0),
-                    bitrate: item.bit_rate.unwrap_or(0).min(u16::MAX as u32) as u16,
-                    track_number: item.track,
-                    disc_number: item.disc_number,
-                    musicbrainz_release_id: None,
-                    musicbrainz_recording_id: None,
-                    musicbrainz_track_id: None,
-                    playlist_item_id: None,
-                    artists: vec![artist],
-                }
-            })
-            .collect())
-    }
-
-    async fn fetch_artist_images(&self) -> Result<Vec<(String, String)>, SourceError> {
-        let artists = self.client.get_artists().await?;
-        let mut out = Vec::new();
-        for artist in artists {
-            if let Some(cover_art_id) = &artist.cover_art
-                && let Ok(url) = self.client.cover_art_url(cover_art_id, Some(512))
-            {
-                out.push((artist.name.clone(), url));
-            }
-        }
-        Ok(out)
-    }
-}
-
-// ========================= YouTube Music ==============================
-
-struct YtSource {
-    db: Db,
-    source: Source,
-    client: YouTubeMusicClient,
-}
-
-#[async_trait]
-impl MediaSource for YtSource {
-    fn source(&self) -> &Source {
-        &self.source
-    }
-    fn db(&self) -> &Db {
-        &self.db
-    }
-
-    fn capabilities(&self) -> Capabilities {
-        // YT has discover + radio, can add/remove playlist entries, but its
-        // InnerTube exposes no reorder mutation — so no `reorder_playlist`
-        // override; it inherits the unsupported default.
-        Capabilities {
-            edit_tags: false,
-            delete_from_disk: false,
-            scan_folders: false,
-            folders: false,
-            sync: true,
-            downloads: true,
-            discover: true,
-            radio: true,
-            playlists: PlaylistOps::AddRemove,
-            artist_view: ArtistView::Remote,
-            albums: AlbumType::YtMusic,
-            favorites_sync: FavoritesSync::Paginated,
-        }
-    }
-
-    async fn start_radio(&self, seed_ref: &str) -> Result<Vec<reader::Track>, SourceError> {
-        if seed_ref.trim().is_empty() {
-            return Err(SourceError::InvalidInput("track has no video id".into()));
-        }
-        // /next works anonymously (empty cookies), so no auth gate here.
-        self.client
-            .start_mix(seed_ref)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    fn web_url(&self, track: &reader::Track) -> Option<String> {
-        let vid = track.id.key();
-        (!vid.trim().is_empty()).then(|| format!("https://music.youtube.com/watch?v={vid}"))
-    }
-
-    async fn search(
-        &self,
-        query: &str,
-    ) -> Result<(Vec<reader::Track>, Vec<reader::Album>), SourceError> {
-        if query.trim().is_empty() {
-            return Ok((Vec::new(), Vec::new()));
-        }
-        let tracks = self.client.search_tracks(query).await?;
-        Ok((tracks, Vec::new()))
-    }
-
-    async fn discover_home(&self) -> Result<crate::ytmusic::discover::DiscoverHome, SourceError> {
-        self.client.discover_home().await.map_err(SourceError::from)
-    }
-
-    async fn discover_continuation(
-        &self,
-        token: &str,
-    ) -> Result<crate::ytmusic::discover::DiscoverHome, SourceError> {
-        self.client
-            .discover_continuation(token)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_album_tracks(&self, browse_id: &str) -> Result<Vec<reader::Track>, SourceError> {
-        self.client
-            .fetch_album_tracks(browse_id)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_album(&self, browse_id: &str) -> Result<RemoteAlbum, SourceError> {
-        self.client
-            .fetch_album(browse_id)
-            .await
-            .map(RemoteAlbum::from)
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_album_by_ref(&self, id: &str) -> Result<Option<RemoteAlbum>, SourceError> {
-        // Resolve the id (raw browse id, `ytmusic:album:MPRE…`, or a synthesized
-        // `ytmusic:album:<hash>`) to a real browse id before fetching.
-        let browse_id = if let Some(bid) = crate::ytmusic::search::album_browse_id(id) {
-            Some(bid)
-        } else if let Some((album, artist)) = crate::ytmusic::search::synth_album_parts(id) {
-            self.resolve_album_browse_id(&album, &artist).await?
-        } else {
-            None
-        };
-        let Some(browse_id) = browse_id else {
-            return Ok(None);
-        };
-        Ok(self
-            .fetch_album(&browse_id)
-            .await
-            .ok()
-            .filter(|a| !a.tracks.is_empty()))
-    }
-
-    async fn fetch_album_by_meta(
-        &self,
-        title: &str,
-        artist: &str,
-    ) -> Result<Option<RemoteAlbum>, SourceError> {
-        let Some(browse_id) = self.resolve_album_browse_id(title, artist).await? else {
-            return Ok(None);
-        };
-        Ok(self
-            .fetch_album(&browse_id)
-            .await
-            .ok()
-            .filter(|a| !a.tracks.is_empty()))
-    }
-
-    async fn fetch_playlist_page(
-        &self,
-        playlist_id: &str,
-        cursor: Option<String>,
-    ) -> Result<(Vec<reader::Track>, Option<String>), SourceError> {
-        self.client
-            .playlist_page(playlist_id, cursor.as_deref())
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn resolve_artist_channel_id(&self, query: &str) -> Result<Option<String>, SourceError> {
-        self.client
-            .resolve_artist_channel_id(query)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn resolve_album_browse_id(
-        &self,
-        album: &str,
-        artist: &str,
-    ) -> Result<Option<String>, SourceError> {
-        self.client
-            .resolve_album_browse_id(album, artist)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_artist(
-        &self,
-        channel_id: &str,
-    ) -> Result<crate::ytmusic::discover::YtArtist, SourceError> {
-        self.client
-            .fetch_artist(channel_id)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn fetch_artist_image(&self, name: &str) -> Result<Option<String>, SourceError> {
-        self.client
-            .resolve_artist_image(name)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn add_to_playlist(
-        &self,
-        playlist_id: &str,
-        item_refs: &[String],
-    ) -> Result<Vec<String>, SourceError> {
-        let mut added = Vec::new();
-        for id in item_refs {
-            if self.client.add_to_playlist(playlist_id, id).await.is_ok() {
-                added.push(id.clone());
-            }
-        }
-        mirror_added(&self.db, &self.source, playlist_id, &added).await?;
-        Ok(added)
-    }
-
-    async fn create_playlist(
-        &self,
-        name: &str,
-        item_refs: &[String],
-    ) -> Result<String, SourceError> {
-        let refs: Vec<&str> = item_refs.iter().map(String::as_str).collect();
-        let id = self.client.create_playlist(name, &refs).await?;
-        mirror_created(&self.db, &self.source, &id, name, item_refs).await?;
-        Ok(id)
-    }
-
-    async fn remove_from_playlist(
-        &self,
-        playlist_id: &str,
-        track: &reader::Track,
-        _position: usize,
-    ) -> Result<(), SourceError> {
-        let vid = track.id.key();
-        if vid.is_empty() {
-            return Err(SourceError::InvalidInput("track has no video id".into()));
-        }
-        self.client.remove_from_playlist(playlist_id, &vid).await?;
-        self.db
-            .remove_playlist_tracks(&self.source, playlist_id, &[vid.into_owned()])
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn resolve_stream(&self, item_id: &str) -> Result<StreamInfo, SourceError> {
-        let info = self.client.get_stream(item_id).await?;
-        Ok(StreamInfo {
-            url: info.url,
-            format: Some((info.format, info.range_safe)),
-            user_agent: Some(info.user_agent),
-            duration_secs: info.duration_secs,
-            bitrate: info.bitrate,
-            content_length: info.content_length,
-        })
-    }
-
-    async fn validate(&self) -> AuthOutcome {
-        match self.client.validate_cookies().await {
-            Ok(()) => AuthOutcome::Valid,
-            Err(e) if e.contains("cookies expired") || e.contains("signed out") => {
-                AuthOutcome::Expired
-            }
-            Err(_) => AuthOutcome::Unreachable,
-        }
-    }
-
-    async fn fetch_favorites(&self) -> Result<Vec<String>, SourceError> {
-        let mut ids = Vec::new();
-        self.client
-            .stream_liked_songs(|page| {
-                ids.extend(page.into_iter().map(|t| t.id.key().into_owned()));
-            })
-            .await?;
-        Ok(ids)
-    }
-
-    async fn push_favorite(&self, item_id: &str, on: bool) -> Result<(), SourceError> {
-        if on {
-            self.client.like_video(item_id).await
-        } else {
-            self.client.unlike_video(item_id).await
-        }
-        .map_err(SourceError::from)
-    }
-
-    async fn fetch_playlists(&self) -> Result<Vec<PlaylistMeta>, SourceError> {
-        Ok(self
-            .client
-            .list_playlists()
-            .await?
-            .into_iter()
-            .map(|s| PlaylistMeta {
-                id: s.id,
-                name: s.title,
-                image_tag: s
-                    .thumbnail_url
-                    .as_ref()
-                    .map(|u| utils::jellyfin_image::encode_cover_url(u)),
-            })
-            .collect())
-    }
-
-    async fn fetch_playlist_entries(
-        &self,
-        playlist_id: &str,
-    ) -> Result<Vec<reader::Track>, SourceError> {
-        // The YT client already returns typed tracks.
-        Ok(self.client.get_playlist_entries(playlist_id).await?)
-    }
-
-    async fn fetch_playlist_entries_page(
-        &self,
-        playlist_id: &str,
-        cursor: Option<String>,
-    ) -> Result<PlaylistPage, SourceError> {
-        // True per-page InnerTube walk so a long playlist streams into the cache
-        // (and the UI) instead of blocking on a full fetch every visit.
-        let (tracks, next) = self
-            .client
-            .playlist_page(playlist_id, cursor.as_deref())
-            .await?;
-        Ok(PlaylistPage { tracks, next })
-    }
-
-    async fn fetch_favorites_page(
-        &self,
-        cursor: Option<String>,
-    ) -> Result<FavoritesPage, SourceError> {
-        let (tracks, next) = self.client.liked_songs_page(cursor.as_deref()).await?;
-        Ok(FavoritesPage { tracks, next })
-    }
 }
 
 // ============================ Resolvers ================================
@@ -1975,32 +762,12 @@ fn active_server_id(config: &AppConfig) -> Option<String> {
 /// Build the per-remote source for `conn` — the ONE place service dispatch happens.
 fn remote_source(db: Db, source: Source, conn: &ServerConn) -> Box<dyn MediaSource> {
     match conn.service {
-        MusicService::Jellyfin => Box::new(JellyfinSource {
-            db,
-            source,
-            client: JellyfinClient::new(
-                &conn.url,
-                Some(&conn.token),
-                &conn.device_id,
-                Some(&conn.user_id),
-            ),
-        }),
-        MusicService::Subsonic | MusicService::Custom => Box::new(SubsonicSource {
-            db,
-            source,
-            client: SubsonicClient::new(&conn.url, &conn.user_id, &conn.token),
-            service: conn.service,
-        }),
-        MusicService::YtMusic => Box::new(YtSource {
-            db,
-            source,
-            client: YouTubeMusicClient::with_cookies(conn.token.clone()),
-        }),
-        MusicService::SoundCloud => Box::new(SoundcloudSource {
-            db,
-            source,
-            token: (!conn.token.is_empty()).then(|| conn.token.clone()),
-        }),
+        MusicService::Jellyfin => Box::new(JellyfinSource::new(db, source, conn)),
+        MusicService::Subsonic | MusicService::Custom => {
+            Box::new(SubsonicSource::new(db, source, conn))
+        }
+        MusicService::YtMusic => Box::new(YtSource::new(db, source, conn)),
+        MusicService::SoundCloud => Box::new(SoundcloudSource::new(db, source, conn)),
     }
 }
 
@@ -2079,172 +846,4 @@ pub fn resolve(db: Db, config: &AppConfig, source: &Source) -> Box<dyn MediaSour
 /// The [`MediaSource`] for the app's active source.
 pub fn active(db: Db, config: &AppConfig) -> Box<dyn MediaSource> {
     resolve(db, config, &config.active_source)
-}
-
-// ============================ SoundCloud ===============================
-
-struct SoundcloudSource {
-    db: Db,
-    source: Source,
-    /// OAuth token for the signed-in account; `None` = anonymous (search + play
-    /// of public tracks still work via the scraped web-player client_id).
-    token: Option<String>,
-}
-
-#[async_trait]
-impl MediaSource for SoundcloudSource {
-    fn source(&self) -> &Source {
-        &self.source
-    }
-    fn db(&self) -> &Db {
-        &self.db
-    }
-
-    fn capabilities(&self) -> Capabilities {
-        Capabilities {
-            edit_tags: false,
-            delete_from_disk: false,
-            scan_folders: false,
-            folders: false,
-            sync: true,
-            downloads: false,
-            discover: false,
-            radio: false,
-            // No write side wired (api-v2 playlist mutation is DataDome-gated).
-            playlists: PlaylistOps::None,
-            artist_view: ArtistView::Library,
-            albums: AlbumType::Standard,
-            favorites_sync: FavoritesSync::Paginated,
-        }
-    }
-
-    async fn resolve_stream(&self, item_id: &str) -> Result<StreamInfo, SourceError> {
-        let url = match crate::soundcloud::resolve_stream(item_id, self.token.as_deref()).await? {
-            // Progressive MP3 streams straight through the normal HTTP path.
-            crate::soundcloud::ResolvedStream::Progressive(u) => u,
-            // HLS (Go+ AAC) is tagged so the player assembles its fMP4 segments
-            // (Symphonia has no HLS demuxer) instead of streaming the .m3u8.
-            crate::soundcloud::ResolvedStream::HlsAac(u) => format!("__SC_HLS:{u}"),
-        };
-        Ok(StreamInfo {
-            url,
-            format: None,
-            user_agent: None,
-            duration_secs: None,
-            bitrate: None,
-            content_length: None,
-        })
-    }
-
-    async fn validate(&self) -> AuthOutcome {
-        match self.token.as_deref() {
-            // Anonymous mode is always usable (public search + play).
-            None => AuthOutcome::Valid,
-            Some(token) => match crate::soundcloud::get_me(token).await {
-                Ok(_) => AuthOutcome::Valid,
-                // Can't cleanly tell expired-token from a network blip, so don't
-                // force a re-sign-in: treat any failure as unreachable.
-                Err(_) => AuthOutcome::Unreachable,
-            },
-        }
-    }
-
-    async fn fetch_favorites(&self) -> Result<Vec<String>, SourceError> {
-        let Some(token) = self.token.as_deref() else {
-            return Ok(Vec::new());
-        };
-        let mut ids = Vec::new();
-        let mut cursor: Option<String> = None;
-        loop {
-            let (tracks, next) =
-                crate::soundcloud::liked_tracks_page(token, cursor.as_deref()).await?;
-            ids.extend(tracks.iter().map(|t| t.id.key().into_owned()));
-            match next {
-                Some(c) => cursor = Some(c),
-                None => break,
-            }
-        }
-        Ok(ids)
-    }
-
-    async fn fetch_favorites_page(
-        &self,
-        cursor: Option<String>,
-    ) -> Result<FavoritesPage, SourceError> {
-        let Some(token) = self.token.as_deref() else {
-            return Ok(FavoritesPage {
-                tracks: Vec::new(),
-                next: None,
-            });
-        };
-        let (tracks, next) = crate::soundcloud::liked_tracks_page(token, cursor.as_deref()).await?;
-        Ok(FavoritesPage { tracks, next })
-    }
-
-    async fn push_favorite(&self, item_id: &str, on: bool) -> Result<(), SourceError> {
-        let token = self.token.as_deref().ok_or(SourceError::Auth)?;
-        crate::soundcloud::set_track_like(item_id, on, token)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn search(
-        &self,
-        query: &str,
-    ) -> Result<(Vec<reader::Track>, Vec<reader::Album>), SourceError> {
-        let tracks = crate::soundcloud::search_tracks(query).await?;
-        Ok((tracks, Vec::new()))
-    }
-
-    async fn fetch_playlists(&self) -> Result<Vec<PlaylistMeta>, SourceError> {
-        let Some(token) = self.token.as_deref() else {
-            return Ok(Vec::new());
-        };
-        Ok(crate::soundcloud::list_playlists(token)
-            .await?
-            .into_iter()
-            .map(|p| PlaylistMeta {
-                id: p.id,
-                name: p.title,
-                image_tag: p.artwork_url,
-            })
-            .collect())
-    }
-
-    async fn fetch_playlist_entries(
-        &self,
-        playlist_id: &str,
-    ) -> Result<Vec<reader::Track>, SourceError> {
-        let Some(token) = self.token.as_deref() else {
-            return Ok(Vec::new());
-        };
-        crate::soundcloud::get_playlist_entries(playlist_id, token)
-            .await
-            .map_err(SourceError::from)
-    }
-
-    async fn add_to_playlist(
-        &self,
-        _playlist_id: &str,
-        _item_refs: &[String],
-    ) -> Result<Vec<String>, SourceError> {
-        Err(SourceError::unsupported("playlist add"))
-    }
-
-    async fn create_playlist(
-        &self,
-        _name: &str,
-        _item_refs: &[String],
-    ) -> Result<String, SourceError> {
-        Err(SourceError::unsupported("playlist create"))
-    }
-
-    async fn remove_from_playlist(
-        &self,
-        _playlist_id: &str,
-        _track: &reader::Track,
-        _position: usize,
-    ) -> Result<(), SourceError> {
-        Err(SourceError::unsupported("playlist remove"))
-    }
 }
